로그인, 회원가입 구현 끝 / 습득물*분실물 게시판 완성(read) 추후에 어떻게 시스팀할지 고민, 게시판 페이징처리 완료, 지도에서 표시할 수 있게 할것

+++ src/main/java/com/lost/site/dto/User.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor //?
 
 public class User {
-	private int user_num;
+	private long user_num;
 	private String user_id;
 	private String user_pw;
 	private String user_email;
